Added simulated chunks that are not rendered yet loaded because some neighbor chunks needs them for computations

+++ code/Chunk/blockLightMod.h
@@ -42,7 +42,7 @@ namespace VoxelEng {
 		* @param intensity Block light modification intensity
 		* @param color Block light modification color
 		*/
-		blockLightMod(const vec3& pos, char intensity, const basicVec4& color);
+		blockLightMod(const ivec3& pos, char intensity, const basicVec4& color);
 
 
 		/*
@@ -52,7 +52,7 @@ namespace VoxelEng {
 		/**
 		* @brief Position the light is being propagated to.
 		*/
-		vec3 pos;
+		ivec3 pos;
 
 		/**
 		* @brief It determines whether the light's color is applied fully in its position or not.
@@ -69,7 +69,7 @@ namespace VoxelEng {
 	inline blockLightMod::blockLightMod() 
 	: pos(vec3Zero), intensity(0), color(basicVec4Zero) {}
 
-	inline blockLightMod::blockLightMod(const vec3& pos, char intensity, const basicVec4& color)
+	inline blockLightMod::blockLightMod(const ivec3& pos, char intensity, const basicVec4& color)
 	: pos(pos), intensity(intensity), color(color) {}
 
 }
@@ -81,11 +81,11 @@ namespace std {
 
 		size_t operator()(const VoxelEng::blockLightMod& t) const {
 		
-			glm::vec3 v;
+			glm::ivec3 v;
 			v.x = t.pos.x;
 			v.y = t.pos.y;
 			v.z = t.pos.z;
-			hash<glm::vec3> h;
+			hash<glm::ivec3> h;
 			return h(v);
 
 		}
